requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
lasio>=0.31
numpy>=1.26
pandas>=2.2
scipy>=1.11
scikit-learn>=1.4
openai>=1.30
google-genai>=1.0.0
